Consolidate Album and Artist image logic into concern
