+++ test/jobs/attach_image_from_discogs_job_test.rb
@@ -16,8 +16,22 @@ class AttachAlbumImageFromDiscogsJobTest < ActiveJob::TestCase
     DiscogsApi.stub(:album_image, "http://example.com/cover.jpg") do
       assert_not album.has_image?
 
-      AttachAlbumImageFromDiscogsJob.perform_now(album)
+      AttachImageFromDiscogsJob.perform_now(album)
       assert album.reload.has_image?
+    end
+  end
+
+  test "should attach image to artist" do
+    artist = artists(:artist1)
+
+    stub_request(:get, "http://example.com/cover.jpg")
+      .to_return(body: file_fixture("cover_image.jpg").read, status: 200, headers: {"Content-Type" => "image/jpeg"})
+
+    DiscogsApi.stub(:artist_image, "http://example.com/cover.jpg") do
+      assert_not artist.has_image?
+
+      AttachImageFromDiscogsJob.perform_now(artist)
+      assert artist.reload.has_image?
     end
   end
 end
